movie-recommender-system/requirements.txt
streamlit==1.22.0
pandas==1.5.3
scikit-learn==1.2.2
imdbpy==2022.7.9
pickle5==0.0.12